start, break apart stats.rs
